+++ usaspending_api/search/tests/integration/hierarchical_filters/test_tas_filter_heirarchical_cases.py
@@ -7,7 +7,7 @@ from usaspending_api.search.tests.integration.hierarchical_filters.tas_fixtures 
     TAS_DICTIONARIES,
     TAS_STRINGS,
 )
-from usaspending_api.search.tests.integration.hierarchical_filters.es_search_test_helpers import (
+from usaspending_api.search.tests.integration.hierarchical_filters.tas_search_test_helpers import (
     _setup_es,
     query_by_tas,
 )
