+++ usaspending_api/search/tests/integration/hierarchical_filters/test_tas_filter_hybrid_cases.py
@@ -5,7 +5,7 @@ from usaspending_api.search.tests.integration.hierarchical_filters.tas_fixtures 
     ATA_TAS,
     TAS_DICTIONARIES,
 )
-from usaspending_api.search.tests.integration.hierarchical_filters.es_search_test_helpers import (
+from usaspending_api.search.tests.integration.hierarchical_filters.tas_search_test_helpers import (
     _setup_es,
     query_by_treasury_account_components,
 )
